feat: please dont panic
Realtime only responded to messages of the websocket while pips are
queried, which lead to broken connections. Also it didn't handled errors
well or logged them in detail, which has been fixed. This changes
the API to listening to a tokio channel and if necessary sending messages,
instead of having control over the actual socket, which is dealt with in
separate threads now.